requirements.txt
slack_bolt
langchain
langchain-community
langchain-openai
python-dotenv